requirements.txt
fastapi==0.92.0
requests==2.28.2
python-dotenv==0.21.1
starlette==0.25.0
uvicorn==0.20.0